+++ sdk/dotnet/ClientGrant.cs
@@ -55,6 +55,14 @@ namespace Pulumi.Auth0
     /// 
     /// }
     /// ```
+    /// 
+    /// ## Import
+    /// 
+    /// Client grants can be imported using the grant ID (Application -&gt; APIs -&gt; Expand the required API)
+    /// 
+    /// ```sh
+    ///  $ pulumi import auth0:index/clientGrant:ClientGrant example cgr_XXXXXXXXXXXXXXXX
+    /// ```
     /// </summary>
     [Auth0ResourceType("auth0:index/clientGrant:ClientGrant")]
     public partial class ClientGrant : Pulumi.CustomResource